deployment_requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
APScheduler==3.10.4
email-validator==2.1.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0